requirements.txt
azure-core==1.29.5
azure-cosmos==4.5.1
azure-functions==1.17.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
